requirements.txt
pycryptodome>=3.19.0
Pillow>=10.0.0
matplotlib>=3.7.2
numpy>=1.24.0